[REFACTOR] Elevado a cobertura de teste.
