+++ src/Components/__tests__/Paper.test.js
@@ -1,11 +1,10 @@
 import React from 'react';
-import Paper from '../Feed/paper';
+import Paper from '../Feed/Paper';
 import { shallow } from 'enzyme';
 
 describe('Testing Paper component', () => {
-    it('Test if paper renders correctly', () =>{
+    it('Test if Paper renders correctly', () =>{
         const tree = shallow(<Paper/>);
         expect(tree).toMatchSnapshot();
     });
 });
-
