+++ src/Components/__tests__/PaperSearch.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import PaperSearch from '../Search/PaperSearch';
+import { shallow } from 'enzyme';
+
+describe('Testing PaperSearch component', () => {
+    it('Test if PaperSearch renders correctly', () =>{
+        const tree = shallow(<PaperSearch/>);
+        expect(tree).toMatchSnapshot();
+    });
+});
